fix: use int32 for CuMesh face tensor in Taubin smoothing
CuMesh.init() expects int32 faces, not int64. This caused a
RuntimeError when Taubin smoothing was enabled.

+++ o-voxel/o_voxel/postprocess.py
@@ -326,7 +326,7 @@ def to_glb(
             iterations=mesh_smooth_iterations,
         )
         smoothed_verts = torch.tensor(tm.vertices, dtype=torch.float32, device=vertices.device)
-        mesh.init(smoothed_verts, torch.tensor(tm.faces, dtype=torch.int64, device=faces.device))
+        mesh.init(smoothed_verts, torch.tensor(tm.faces, dtype=torch.int32, device=faces.device))
         if verbose:
             print(f"After smoothing: {mesh.num_vertices} vertices, {mesh.num_faces} faces")
